python/pyproject.toml
[tool.poetry]
name = "nats-py-tg"
version = "0.1.0"
description = ""
authors = ["Thomas Zub <thomas.zub@outlook.de>"]
readme = "README.md"
package-mode = false

[package]
include = ["src/*.py"]

[tool.poetry.dependencies]
python = "^3.12"
nats-py = "^2.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
